docker/janus-pro/requirements.txt
# Janus-Pro Service Requirements
# For Visual Chart Pattern Recognition

# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiohttp>=3.9.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Hugging Face for Janus-Pro model
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.25.0

# For SigLIP encoder (used by Janus-Pro)
timm>=0.9.0
